requirements.txt
accelerate==0.34.2
transformers==4.45.1
colorama==0.4.6
datasets==2.15.0
diffusers==0.25.0
lpips==0.1.4
matplotlib==3.8.1
numpy==1.26.0
opencv_python==3.4.18.65
opencv_python_headless==4.5.5.64
pandas==2.1.1
Pillow==11.0.0
safetensors==0.4.5
scipy==1.11.1
tensorboardX==2.6.2.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
tqdm==4.66.1
huggingface_hub==0.25.1